+++ bootstrap/bootstrap-api/pom.xml
@@ -40,10 +40,6 @@
         </dependency>
         <dependency>
             <groupId>com.alipay.sofa</groupId>
-            <artifactId>sofa-rpc-codec-jackson</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.alipay.sofa</groupId>
             <artifactId>sofa-rpc-filter</artifactId>
         </dependency>
         <dependency>
