+++ bootstrap/bootstrap-api/src/main/java/com/alipay/sofa/rpc/bootstrap/DefaultProviderBootstrap.java
@@ -16,7 +16,7 @@
  */
 package com.alipay.sofa.rpc.bootstrap;
 
-import com.alipay.sofa.rpc.codec.jackson.JacksonSerializer;
+import com.alipay.sofa.rpc.codec.AbstractSerializer;
 import com.alipay.sofa.rpc.common.struct.NamedThreadFactory;
 import com.alipay.sofa.rpc.common.utils.CommonUtils;
 import com.alipay.sofa.rpc.common.utils.StringUtils;
@@ -190,7 +190,7 @@ public class DefaultProviderBootstrap<T> extends ProviderBootstrap<T> {
             //如果是泛型接口则需要在JSON序列化器中注册跟真实实现类的对应关系，因为反序列化时无法从泛型接口中拿到真实的数据类型
             if (providerConfig.getProxyClass().getTypeParameters().length > 0) {
                 String serviceName = ConfigUniqueNameGenerator.getUniqueName(providerConfig);
-                JacksonSerializer.registerGenericService(serviceName, providerConfig.getRef().getClass().getName());
+                AbstractSerializer.registerGenericService(serviceName, providerConfig.getRef().getClass().getName());
             }
 
             // 注册到注册中心
